+++ tests/testthat/test-verify-output.R
@@ -77,3 +77,9 @@ test_that("verify_output() doesn't use cli unicode by default", {
       cat(cli::symbol$info, cli::symbol$cross, "\n")
     })
 })
+
+test_that("verify_output() handles carriage return", {
+  verify_output(test_path("test-verify-conditions-cr.txt"), {
+    cat("\r\n")
+  })
+})