+++ NEWS.md
@@ -1,5 +1,8 @@
 # testthat (development version)
 
+* `verify_output()` no longer always fails if output contains a carriage return character ("\r")
+  (#1048).
+
 * `verify_output()` now sets `options(useFancyQuotes = FALSE, outDec = ".")`  
   to increase reproducibility across systems (#1000). It also sets
   `options(lifecycle_verbosity = "warning")` to ensure that lifecycle